create MapComponent component
