+++ client/src/App.jsx
@@ -6,87 +6,7 @@ import {
   Autocomplete,
 } from "@react-google-maps/api";
 
-const containerStyle = {
-  width: "400px",
-  height: "400px",
-};
-
-const center = {
-  lat: 43.6532,
-  lng: -79.3832,
-};
-
-const libraries = ["places"];
-
-function MapComponent() {
-  const { isLoaded } = useJsApiLoader({
-    id: "google-map-script",
-    googleMapsApiKey: "AIzaSyBC-KOT7TR068hqqTnswn4xaLBNA79hETs",
-    libraries: libraries,
-  });
-
-  const [map, setMap] = React.useState(null);
-
-  const onUnmount = React.useCallback(function callback(map) {
-    setMap(null);
-  }, []);
-
-  return isLoaded ? (
-    <GoogleMap
-      mapContainerStyle={containerStyle}
-      center={center}
-      zoom={10}
-      onUnmount={onUnmount}
-    >
-      <Autocomplete>
-        <input
-          type="text"
-          placeholder="Origin"
-          style={{
-            boxSizing: `border-box`,
-            border: `1px solid transparent`,
-            width: `240px`,
-            height: `32px`,
-            padding: `0 12px`,
-            borderRadius: `3px`,
-            boxShadow: `0 2px 6px rgba(0, 0, 0, 0.3)`,
-            fontSize: `14px`,
-            outline: `none`,
-            textOverflow: `ellipses`,
-            position: "absolute",
-            left: "50%",
-            marginLeft: "-120px",
-          }}
-        />
-      </Autocomplete>
-      <Autocomplete>
-        <input
-          type="text"
-          placeholder="Destination"
-          style={{
-            boxSizing: `border-box`,
-            border: `1px solid transparent`,
-            width: `240px`,
-            height: `32px`,
-            padding: `0 12px`,
-            borderRadius: `3px`,
-            boxShadow: `0 2px 6px rgba(0, 0, 0, 0.3)`,
-            fontSize: `14px`,
-            outline: `none`,
-            textOverflow: `ellipses`,
-            position: "absolute",
-            left: "50%",
-            top: `32px`,
-            marginLeft: "-120px",
-          }}
-        />
-      </Autocomplete>
-      <Marker position={center} />
-    </GoogleMap>
-  ) : (
-    <></>
-  );
-}
+import MapComponent from "./components/MapComponent";
 
 function App() {
   return (
